handle delete list and todo errors

+++ src/controllers/list/DeleteListController.js
@@ -1,4 +1,5 @@
 import List from "../../models/List.js";
+import Todo from "../../models/Todo.js"; 
 import validId from "../../utils/validId.js";
 
 const deleteList = async (req, res, next) => {
@@ -7,12 +8,24 @@ const deleteList = async (req, res, next) => {
 
         validId([listId]);
 
-        const deleteList = await List.findByIdAndDelete(listId);
-        
-        if(!deleteList) {
+
+        const findList = await List.findById(listId);
+        if (!findList) {
             return res.status(404).json({message: "List not found"});
         }
+
+        const todos = await Todo.find({listId: listId});
+        if (!todos) {
+            return res.status(404).json({message: "List doesn't have todos"});
+        }
+
+        todos.forEach(async (todo) => {
+            await Todo.findByIdAndDelete(todo._id);
+        })
+
+        await List.findByIdAndDelete(listId);
         res.status(200).json({message: "List deleted successfully"});
+
     } catch (error) {
         next(error);
     }
