+++ src/controllers/todo/DeleteTodoController.js
@@ -1,4 +1,5 @@
 import Todo from "../../models/Todo.js";
+import List from "../../models/List.js";
 import validId from "../../utils/validId.js";
 
 const deleteTodo = async (req,res,next) => {
@@ -7,10 +8,20 @@ const deleteTodo = async (req,res,next) => {
         
         validId([todoId]);
 
-        const deleteTodo = await Todo.findByIdAndDelete(todoId);
-        if (!deleteTodo) {
+        const findTodo = await Todo.findById(todoId);
+        if (!findTodo) {
             return res.status(404).json({message: "Todo not found"});
         }
+        
+        const listId = findTodo.listId;
+        const list = await List.findById(listId);
+        if (!list) {
+            return res.status(404).json({message: "List not found"});
+        }
+        list.todos.pop(todoId);
+        await list.save();
+
+        await Todo.findByIdAndDelete(todoId);
         return res.json({message: "Todo deleted successfully"});
